requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
charset-normalizer==3.3.2
urllib3==2.2.3
PyYAML==6.0.2
gspread==6.1.2
google-auth==2.35.0
